1.修改kernel.hpp中Jsize的类型,修改为size_t。 2.实现远程切机模块。 3.实现更确切的错误返回。 4.修复Sensor检测模块和Sensor触发日志查询模块中线路的值在位或的情况下会出现异常线路值。 5.新增系统文件操作模块smart_resouces_android_files_operator.hpp,这里将用来实现系统文件的增删查改,并且具备指令模式调用。

+++ kernel.hpp
@@ -9,7 +9,7 @@
 
 #include <stdint.h>
 
-using Jsize = uint32_t;
+using Jsize = size_t;
 using Jbool = bool;
 using Jbyte = uint8_t;
 using Juint = uint32_t;
